+++ form-view-model/src/main/kotlin/com/ritense/formviewmodel/error/FormErrorsException.kt
@@ -14,10 +14,13 @@
  * limitations under the License.
  */
 
-package com.ritense.formviewmodel.web.rest.dto
+package com.ritense.formviewmodel.error
 
-import com.ritense.formviewmodel.error.MultipleFormException
-
-class MultipleFormError (
-    val componentErrors: List<MultipleFormException.ComponentError> = emptyList(),
-)
+class FormErrorsException(
+    val componentErrors: List<ComponentError> = emptyList()
+) : Exception() {
+    data class ComponentError(
+        val component: String?,
+        val message: String
+    )
+}
